Many updates and bugfixes. Deleted unmantained jobs

+++ setup.sh
@@ -35,17 +35,18 @@ prepare_copy() {
 # Install Debian build tools
 install_debian_buildtools() {
     apt-get install -y \
-        unzip wget git \
+        unzip p7zip-full wget git \
         build-essential debhelper apt-utils fakeroot cmake \
-        python-all-dev python3-all-dev \
+        cargo rustc \
+        python3-all-dev \
         pkg-config libssl-dev autotools-dev zlib1g-dev
 }
 
 # Install Debian useful tools
 install_debian_utils() {
     apt-get install -y \
-        sudo curl vim less procps \
-        silversearcher-ag fd-find tree \
+        sudo curl vim less procps jq \
+        ripgrep tree \
         p7zip bzip2 libbz2-dev \
         gnupg dirmngr
 }
@@ -53,26 +54,24 @@ install_debian_utils() {
 # Install rvt2 dependencies available in Debian
 install_debian_deps() {
     apt-get install -y \
-        sleuthkit \
-        bindfs dislocker ewf-tools testdisk \
-        volatility volatility-tools \
+        bindfs dislocker ewf-tools libewf-dev testdisk \
         libimage-exiftool-perl \
-        libscca1 libscca-utils python3-libscca \
-        libmsiecf1 libmsiecf-utils python3-libmsiecf \
-        liblnk1 liblnk-utils python3-liblnk \
-        fuse3 libfuse3-dev libicu-dev libattr1-dev
+        fuse3 libfuse3-dev libfuse-dev libicu-dev libattr1-dev
     apt-get install -y libparse-win32registry-perl # regripper
 }
 
 # Install python3 and pip3
 install_debian_python() {
-    apt-get install -y python3 python3-pip python-setuptools python3-setuptools
+    apt-get install -y python3 python3-pip python3-setuptools
 }
 
 # Install rvt2 python dependencies
 install_pip_deps() (
-    pip3 install lz4
-    pip3 install pipenv
+    # these dependencies fail in the Dockerfile
+    python3 -m pip install lz4
+    python3 -m pip install evtx
+    python3 -m pip install pipenv
+    python3 -m pip install pycrypto
     export LC_ALL=C.UTF-8
     export LANG=C.UTF-8
     cd "${RVT2HOME}"
@@ -80,11 +79,22 @@ install_pip_deps() (
     pipenv install
 )
 
+build_install_evtx() (
+    cd "${SRCDIR}"
+    EVTX=$(curl --silent "https://api.github.com/repos/omerbenamram/evtx/releases/latest"| grep -oP '"browser_download_url": "\K(.*)(?=")' |grep linux-gnu| grep x86_64)
+    wget $EVTX
+    chmod 755 evtx_dump*
+    mv evtx_dump* /usr/local/bin/evtx_dump
+)
+
 build_install_sleuthkit() (
     cd "${SRCDIR}"
-    git clone "https://github.com/sleuthkit/sleuthkit.git"
-    cd sleuthkit
-    ./bootstrap
+    SLEUTHKIT=$(curl --silent "https://api.github.com/repos/sleuthkit/sleuthkit/releases/latest" | grep -oP '"browser_download_url": "\K(.*)(?=")'| grep "tar.gz$")
+    wget $SLEUTHKIT
+    FILENAME=${SLEUTHKIT##*/}
+    tar xzvf $FILENAME
+    rm $FILENAME
+    cd "${FILENAME%.tar.gz}"
     ./configure
     make
     make install
@@ -155,31 +165,50 @@ _build_install_libyal() (
     ln -s ${NAME}-${STAGE}-${VER}.tar.gz ${NAME}_${VER}.orig.tar.gz
     tar xzf ${NAME}-${STAGE}-${VER}.tar.gz
     cd ${NAME}-${VER}/
-    cp -rf dpkg debian
-    dpkg-buildpackage -uc -us -rfakeroot
+    ./configure --enable-python
+    make
+    make install
+    ldconfig
     cd ..
-    dpkg -i ${NAME}_*.deb ${NAME}-tools_*.deb ${NAME}-python3_*.deb
 )
 
 build_install_libesedb() {
-    local VERSION=$(curl -s "https://github.com/libyal/libesedb/releases" \
-    | grep "libesedb-experimental-.*.tar.gz" | head -1 \
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libesedb/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
     | sed -rn "s/.*([0-9]{8}).*/\1/p")
 
     _build_install_libyal libesedb experimental "${VERSION}"
 }
 
-build_install_libpff() {
-    local VERSION=$(curl -s "https://github.com/libyal/libpff/releases" \
-    | grep "libpff-experimental-.*.tar.gz" | head -1 \
+build_install_liblnk() {
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/liblnk/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
     | sed -rn "s/.*([0-9]{8}).*/\1/p")
 
-    _build_install_libyal libpff experimental "${VERSION}"
+    _build_install_libyal liblnk alpha "${VERSION}"
+}
+
+build_install_libmsiecf() {
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libmsiecf/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
+    | sed -rn "s/.*([0-9]{8}).*/\1/p")
+
+    _build_install_libyal libmsiecf alpha "${VERSION}"
+}
+
+build_install_libscca() {
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libscca/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
+    | sed -rn "s/.*([0-9]{8}).*/\1/p")
+
+    _build_install_libyal libscca alpha "${VERSION}"
+}
+
+build_install_libpff() {
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libpff/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
+    | sed -rn "s/.*([0-9]{8}).*/\1/p")
+
+    _build_install_libyal libpff alpha "${VERSION}"
 }
 
 build_install_libvshadow() {
-    local VERSION=$(curl -s "https://github.com/libyal/libvshadow/releases" \
-    | grep "libvshadow-alpha-.*.tar.gz" | head -1 \
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libvshadow/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
     | sed -rn "s/.*([0-9]{8}).*/\1/p")
 
     _build_install_libyal libvshadow alpha "${VERSION}"
@@ -187,21 +216,37 @@ build_install_libvshadow() {
 }
 
 build_install_libfvde() {
-    local VERSION=$(curl -s "https://github.com/libyal/libfvde/releases" \
-    | grep "libfvde-experimental-.*.tar.gz" | head -1 \
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libfvde/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
     | sed -rn "s/.*([0-9]{8}).*/\1/p")
 
     _build_install_libyal libfvde experimental "${VERSION}"
     # sed -i "s/.user_allow_other/user_allow_other/" /etc/fuser.conf
 }
 
+build_install_libevt() {
+    local VERSION=$(curl -s "https://api.github.com/repos/libyal/libevt/releases" | grep -oP '"browser_download_url":\s*"\K(.*)(?=")'| grep "tar.gz$" | head -1 \
+    | sed -rn "s/.*([0-9]{8}).*/\1/p")
+
+    _build_install_libyal libevt alpha "${VERSION}"
+}
+
+build_install_volatility() {
+    cd "${SRCDIR}"
+    git clone https://github.com/volatilityfoundation/volatility3.git
+    ln -s "$(pwd)"/volatility3/vol.py /usr/local/bin/vol.py
+    mkdir -p volatility3/symbols
+    wget -O volatility3/symbols/windows.zip https://downloads.volatilityfoundation.org/volatility3/symbols/windows.zip
+    wget -O volatility3/symbols/mac.zip https://downloads.volatilityfoundation.org/volatility3/symbols/mac.zip
+    wget -O volatility3/symbols/linux.zip https://downloads.volatilityfoundation.org/volatility3/symbols/linux.zip
+}
+
 build_install_regripper() (
-    local NAME="RegRipper2.8"
-    local COMMIT="ee874d5245fb4f26147c29dc1db02b8e68a88698"
+    local NAME="RegRipper3.0"
+    local COMMIT="c59d23d151059aa98f1888ee52e2eea16f746f21"
     [ -d /tmp/patches ] || cp -rp patches /tmp/
     cd "${SRCDIR}"
     _download_verify "https://github.com/keydet89/${NAME}/archive/${COMMIT}.zip" \
-        "9cea8786588417b89a6f9497d8d97222f5f7daeaf276b40a2cd02157ea121b2e"
+        "db7040058e80fe816e2a8936a29c5a62ebab235eec0870d306453832a0343335"
     mv "${COMMIT}".zip ${NAME}-${COMMIT}.zip
     unzip ${NAME}-${COMMIT}.zip
     cd ${NAME}-${COMMIT}
@@ -209,11 +254,33 @@ build_install_regripper() (
     patch rip.pl < /tmp/patches/regripper_patch.diff
 
     install -p -dm 755 /opt/regripper/
-    install -p -m 755 *.pl *.txt *.md *.pdf /opt/regripper/
+    install -p -m 755 *.pl *.txt *.md /opt/regripper/
     install -p -dm 755 /opt/regripper/plugins/
     install -p -m 755 plugins/* /opt/regripper/plugins/
-    install -p -m 755 /tmp/patches/regripper_plugins/outlook_search.pl \
-	      /tmp/patches/regripper_plugins/winlogon_db.pl \
+    install -p -m 755 /tmp/patches/regripper_plugins/ccleaner.pl \
+          /tmp/patches/regripper_plugins/cortana.pl \
+          /tmp/patches/regripper_plugins/defbrowser.pl \
+          /tmp/patches/regripper_plugins/diag_sr.pl \
+          /tmp/patches/regripper_plugins/eventlog.pl \
+          /tmp/patches/regripper_plugins/eventlogs.pl \
+          /tmp/patches/regripper_plugins/fw_config.pl \
+          /tmp/patches/regripper_plugins/ie_settings.pl \
+          /tmp/patches/regripper_plugins/ie_version.pl \
+          /tmp/patches/regripper_plugins/outlook_search.pl \
+          /tmp/patches/regripper_plugins/polacdms.pl \
+          /tmp/patches/regripper_plugins/proxysettings.pl \
+          /tmp/patches/regripper_plugins/rdphint.pl \
+          /tmp/patches/regripper_plugins/rdpnla.pl \
+          /tmp/patches/regripper_plugins/silentprocessexit.pl \
+          /tmp/patches/regripper_plugins/silentprocessexit_tln.pl \
+          /tmp/patches/regripper_plugins/teamviewer.pl \
+          /tmp/patches/regripper_plugins/vmware_vsphere_client.pl \
+          /tmp/patches/regripper_plugins/winevt.pl \
+          /tmp/patches/regripper_plugins/winlogon_db.pl \
+          /tmp/patches/regripper_plugins/winnt_cv.pl \
+          /tmp/patches/regripper_plugins/winscp_sessions.pl \
+          /tmp/patches/regripper_plugins/crashcontrol.pl \
+          /tmp/patches/regripper_plugins/winver2.pl \
         /opt/regripper/plugins/
     install -p -dm 755 /usr/local/bin/
     ln -s /opt/regripper/rip.pl /usr/local/bin/rip
@@ -222,12 +289,18 @@ build_install_regripper() (
     # Replaces Base, File and Key.pm with regripper version
 
     local WIN32REGISTRY=$(dirname $(dpkg -L libparse-win32registry-perl|grep "/Base.pm"))
-    sudo cp -np "$WIN32REGISTRY/Base.pm" "$WIN32REGISTRY/Base.pm.old"
-    sudo cp -p Base.pm "$WIN32REGISTRY"
-    sudo cp -np "$WIN32REGISTRY/WinNT/File.pm" "$WIN32REGISTRY/WinNT/File.pm.old"
-    sudo cp -p File.pm "$WIN32REGISTRY/WinNT/"
-    sudo cp -np "$WIN32REGISTRY/WinNT/Key.pm" "$WIN32REGISTRY/WinNT/Key.pm.old"
-    sudo cp -p Key.pm "$WIN32REGISTRY/WinNT/"
+    if [ -e Base.pm ]; then
+        sudo cp -np "$WIN32REGISTRY/Base.pm" "$WIN32REGISTRY/Base.pm.old"
+        sudo cp -p Base.pm "$WIN32REGISTRY"
+    fi
+    if [ -e File.pm ]; then
+        sudo cp -np "$WIN32REGISTRY/WinNT/File.pm" "$WIN32REGISTRY/WinNT/File.pm.old"
+        sudo cp -p File.pm "$WIN32REGISTRY/WinNT/"
+    fi
+    if [ -e Key.pm ]; then
+        sudo cp -np "$WIN32REGISTRY/WinNT/Key.pm" "$WIN32REGISTRY/WinNT/Key.pm.old"
+        sudo cp -p Key.pm "$WIN32REGISTRY/WinNT/"
+    fi
 )
 
 build_install_ntfs3g_system_compression() (
@@ -261,26 +334,19 @@ build_install_apfs_fuse() (
 
 build_install_yara() (
     cd "${SRCDIR}"
-    wget https://github.com/VirusTotal/yara/archive/v3.11.0.tar.gz
-    tar -zxf v3.11.0.tar.gz
-    cd yara-3.11.0
+    YARA=$(curl --silent "https://api.github.com/repos/VirusTotal/yara/releases/latest" | grep -oP '"tarball_url": "\K(.*)(?=")')
+    wget $YARA
+    FILENAME=${YARA##*/}
+    tar xzvf $FILENAME
+    rm $FILENAME
+    mv VirusTotal-yara-* yara-$FILENAME
+    cd yara-$FILENAME
+
     ./bootstrap.sh
-    apt-get install -y flex bison
+    apt-get install -y flex bison libjansson-dev libmagic-dev
     ./configure --with-crypto --enable-cuckoo --enable-magic --enable-dotnet
     make
     make install
-)
-
-build_install_hindsight() (
-    cd ${SRCDIR}
-    git clone --depth 1 https://github.com/obsidianforensics/hindsight.git
-    cd hindsight
-    apt-get install python-pip
-    yes | pip install -r requirements.txt  # segment violation during protobuf
-    # pip install protobuf scipy xlsxwriter bottle # if segment error
-    chmod 775 ${SRCDIR}/hindsight/hindsight.py  # ???
-    # make a symbolic link to /usr/local/bin ???
-    ln -s ${SRCDIR}/hindsight/hindsight.py /usr/local/bin/hindsight.py
 )
 
 # Clone submodules and make some patches to submodules in order to work as RVT2 expects
@@ -290,6 +356,37 @@ submodules() (
     # srumpdump includes a graphical interface not needed in RVT2
     sed '19d' plugins/external/srum-dump/srum_dump2.py > tempfile && mv tempfile plugins/external/srum-dump/srum_dump2.py
 )
+
+install_zimmerman_tools(){
+# dotnet installation
+    cd "${RVT2HOME}"
+    EXTERNAL_PATH="external_tools"
+    mkdir -p $EXTERNAL_PATH/dotnet
+    mkdir $EXTERNAL_PATH/windows
+    wget https://dot.net/v1/dotnet-install.sh -O dotnet-install.sh
+    chmod +x ./dotnet-install.sh 
+    ./dotnet-install.sh -c 6.0 --runtime dotnet -i $EXTERNAL_PATH/dotnet
+    rm ./dotnet-install.sh
+    # Zimmerman tools
+    cd $EXTERNAL_PATH
+    cd windows
+    for tool in "AmcacheParser" "AppCompatCacheParser" "MFTECmd" "SBECmd" "SrumECmd" "WxTCmd"
+      do
+        mkdir $tool
+        cd $tool
+        wget https://f001.backblazeb2.com/file/EricZimmermanTools/net6/${tool}.zip
+        7z x $tool.zip
+        rm $tool.zip
+        cd ..
+      done
+    for tool in "SDBExplorer" "RECmd"
+      do
+        wget https://f001.backblazeb2.com/file/EricZimmermanTools/net6/${tool}.zip
+        7z x $tool.zip
+        rm $tool.zip
+      done
+    cd ..
+}
 
 install_rvt_bin() {
     cat << EOF > /usr/local/bin/rvt2
@@ -322,12 +419,13 @@ create_link_bin() {
 # Remove unnecessary source files
 clean_sources() (
   cd "${SRCDIR}"
-  rm -rf libvshadow* libesedb* libpff* libfvde* apfs-fuse ntfs-3g*
+  rm -rf libvshadow* libesedb* libpff* libfvde* libscca* libevt* apfs-fuse ntfs-3g* liblnk libmsiecf
   rm -f Parse-Evtx*.zip* RegRipper2*
 )
 
-# Use this function to install all dependencies on Debian stretch and setup rvt2
+# Use this function to install all dependencies on ubuntu 22.04 and setup rvt2
 setup_debian_full() {
+
     # Basic preparation
     prepare_debian
     prepare
@@ -339,24 +437,31 @@ setup_debian_full() {
     install_debian_deps
     install_debian_python
 
+    # Download submodules
+    submodules
+
+    install_zimmerman_tools
+
     # Extra Tools
-      # build_install_sleuthkit
+    build_install_sleuthkit
       # build_install_sleuthkit_APFS
     build_install_libesedb
     build_install_libpff
     build_install_libvshadow
     build_install_libfvde
+    build_install_liblnk
+    build_install_libmsiecf
+    build_install_libscca
+    build_install_libevt
     build_install_regripper
-    build_install_ntfs3g_system_compression
+    # build_install_ntfs3g_system_compression
     build_install_apfs_fuse
     build_install_yara
-      # build_install_hindsight
+    build_install_volatility
+    build_install_evtx
 
     # Install pip dependencies
     install_pip_deps
-
-    # Download submodules
-    submodules
 
     # Permissions and links
     prepare_sudo
@@ -372,6 +477,6 @@ build_install_evtxparser() {
   :
 }
 
-if [ $1 = "run" ]; then
+if [ "$1" = "run" ]; then
   setup_debian_full
 fi
